django_project/requirements.txt
Django==4.1.7
djangorestframework==3.14.0
django-cors-headers==3.13.0
psycopg2-binary==2.8.6
django-phonenumber-field==7.0.2
gunicorn==20.1.0